% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functionality_summary.R
\name{data_functionality_summary}
\alias{data_functionality_summary}
\title{Summarize the Functionality in Group, Batch and DEGs of A Simulation Method}
\usage{
data_functionality_summary(
  data,
  group = NULL,
  batch = NULL,
  k = NULL,
  DEGs = NULL,
  DEA_method = "edgeRQLFDetRate",
  model_method = "SVM",
  ref_data = NULL,
  ref_data_grouping = NULL,
  algorithm = "Hungarian",
  seed = 1,
  verbose = TRUE,
  threads = 1
)
}
\arguments{
\item{data}{The matrix or data frame of gene expression profile or the output by \code{simpipe::simulate_datasets}.}

\item{group}{The characters of group names which cells belong to. Needed if data is a matrix or a dataframe and you want to evaluate the groups, DEGs and trajectories.}

\item{batch}{The characters of batch names which cells belong to. Needed if data is a matrix or a dataframe and you want to evaluate the batches}

\item{k}{k-nearest neighborhoods of cells.}

\item{DEGs}{A list of DEGs with the names of \code{xxxvsxxx}. Note that the names of DEGs are in the rownames of the matrix or the dataframe and the names of \code{xxx} is in the \code{group} characters. Needed if data is a matrix or a dataframe and you want to evaluate the DEGs.}

\item{DEA_method}{The DEA method to get the DEGs. Choices: edgeRQLF, edgeRQLFDetRate, MASTcpmDetRate, MASTtpmDetRate, MASTcpm, MASTtpm, limmatrend, limmavoom, ttest and wilcox. Default is edgeRQLFDetRate.}

\item{model_method}{The method to establish the model. SVM, Decision tree or RF (Random Forest). Default is SVM.}

\item{ref_data}{A matrix, data frame or list of real gene expression profiles.}

\item{ref_data_grouping}{The vector or list of group names which real cells belong to.}

\item{algorithm}{The algorithm for matching the real cells and the simulated cells. Hungarian (default) or Improved_Hungarian.}

\item{seed}{Random seed for trajectory inference.}

\item{verbose}{Whether the process massages are returned.}

\item{threads}{How many cores used for parallel computation.}
}
\value{
A list of three aspects of the data
}
\description{
Summarize the Functionality in Group, Batch and DEGs of A Simulation Method
}
