% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_properties_summary.R
\name{data_properties_summary}
\alias{data_properties_summary}
\title{Summary Data Properties}
\usage{
data_properties_summary(
  ref_data_cell_properties,
  sim_data_cell_properties,
  ref_data_gene_properties,
  sim_data_gene_properties
)
}
\arguments{
\item{ref_data_cell_properties}{Cell properties of reference data}

\item{sim_data_cell_properties}{Cell properties of simulated data}

\item{ref_data_gene_properties}{Gene properties of reference data}

\item{sim_data_gene_properties}{Gene properties of simulated data}
}
\value{
A list
}
\description{
Summary Data Properties
}
