% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate parameters from Real Datasets}
\usage{
estimate_parameters(
  ref_data,
  method,
  other_prior = list(),
  seed = simutils::random_seed(),
  verbose = TRUE,
  use_docker = FALSE
)
}
\arguments{
\item{ref_data}{A matrix for one dataset or a list of datasets with their own
names.}

\item{method}{A character or a string of methods. It can be \code{all} as all
methods are used to estimate paramters from real data.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them.}

\item{seed}{A random seed. Default value is generated by \code{\link[simutils]{random_seed}}.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{use_docker}{Logical. Default is FALSE. Whether to execute the estimation
step by establishing a Docker container or not. Suggest that you would better
use Docker to execute the step in order to avoid the issues caused by coding
environment.}
}
\value{
A list of estimated parameters and detection results
}
\description{
This is the core function of estimating parameters from real single-cell RNA
sequencing datasets. Users can get many estimation results by pointing out
more than two methods and inputting many real datasets.
}
