% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_DEGs_list}
\alias{prepare_DEGs_list}
\title{Prepare the list of DEGs of each group pair}
\usage{
prepare_DEGs_list(col_meta, row_meta)
}
\arguments{
\item{col_meta}{cell information}

\item{row_meta}{gene information}
}
\description{
Prepare the list of DEGs of each group pair
}
