% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_datasets.R
\name{simulate_datasets}
\alias{simulate_datasets}
\title{Simulate scRNA-seq Datasets}
\usage{
simulate_datasets(
  method = NULL,
  parameters = NULL,
  ref_data = NULL,
  other_prior = list(),
  n = 1,
  seed = simutils::random_seed(),
  return_format = "SingleCellExperiment",
  verbose = TRUE,
  use_docker = FALSE
)
}
\arguments{
\item{method}{A character or a string of methods or NULL. It can be \code{all}
as all methods are used to estimate paramters from real data. If NULL, the
argument \code{parameters} must be generated by \code{\link[=estimate_parameters]{estimate_parameters()}}.}

\item{parameters}{NULL or the object generated by \code{\link[=estimate_parameters]{estimate_parameters()}}.
If NULL, you must input the argument \code{method} above and the default
parameters will be used to perform the simulation step.}

\item{ref_data}{A matrix for one dataset or a list of datasets with their own
names. This is usually unused except for some methods, e.g. SCRIP.}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}

\item{n}{The time(s) that users want to perform the simulation step with every
method.}

\item{seed}{A random seed. An integer or a vector of integers. If you want to
simulate twice or more times by every method you have chosen, it should be a
vector. If you do not input this parameter, the default value(s) will be used.}

\item{return_format}{A character. Alternatives choices: list, SingleCellExperiment,
Seurat and h5ad.}

\item{verbose}{Logical. Whether to return messages or not.}

\item{use_docker}{Logical. Default is FALSE. Whether to execute the simulation
step by establishing a Docker container or not. Suggest that you would better
use Docker to execute the step in order to avoid the issues caused by coding
environment.}
}
\description{
This is the core function of simulating scRNA-seq datasets. Users can execute
this step without parameters estimation step as we will set some default parameters.
Otherwise, we provide an alternative choice to complete this step via establishing
Docker containers.
}
