% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functionality_summary.R
\name{data_functionality_summary}
\alias{data_functionality_summary}
\title{Summary}
\usage{
data_functionality_summary(
  data,
  group = NULL,
  batch = NULL,
  k = NULL,
  DEGs = NULL,
  verbose = NULL,
  threads = 1
)
}
\arguments{
\item{data}{The matrix or data frame of gene expression profile or the output by \code{simpipe::simulate_datasets}.}

\item{group}{The characters of group names which cells belong to. Needed if data is a matrix or a dataframe and you want to evaluate the groups.}

\item{batch}{The characters of batch names which cells belong to. Needed if data is a matrix or a dataframe and you want to evaluate the batches}

\item{k}{k-nearest neighborhoods of cells.}

\item{DEGs}{A list of DEGs with the names of \code{xxxvsxxx}. Note that the names of DEGs are in the rownames of the matrix or the dataframe and the names of \code{xxx} is in the \code{group} characters. Needed if data is a matrix or a dataframe and you want to evaluate the DEGs.}

\item{verbose}{Whether the process massages are returned.}

\item{threads}{How many cores used for parallel computation}
}
\value{
A list of three aspects of the data
}
\description{
Summary
}
